+++ src/java/org/apache/commons/math/stat/univariate/rank/Max.java
@@ -1,7 +1,7 @@
 /* ====================================================================
  * The Apache Software License, Version 1.1
  *
- * Copyright (c) 2003 The Apache Software Foundation.  All rights
+ * Copyright (c) 2003-2004 The Apache Software Foundation.  All rights
  * reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -64,14 +64,14 @@ import org
     .AbstractStorelessUnivariateStatistic;
 
 /**
- * @version $Revision: 1.12 $ $Date: 2003/11/19 03:28:24 $
+ * @version $Revision: 1.13 $ $Date: 2004/01/25 21:30:41 $
  */
 public class Max extends AbstractStorelessUnivariateStatistic implements Serializable {
 
     static final long serialVersionUID = -5593383832225844641L;    
     
     /** */
-    private int n = 0;
+    private long n = 0;
         
     /** */
     private double value = Double.NaN;
